Bugfix: setIndeterminate creates missing drawable
As it stands, if you fail to pass in `mpb_setBothDrawables` and try to
use the method `setIndeterminate(boolean)`, the progress bar will either
disappear or revert to the system drawable. This PR creates the missing
drawable if the `setIndeterminate(boolean)` method is called and the
resulting drawable isn't one of the custom drawables.